requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.7
django-filter==2.4.0
django-utils-six==2.0
djangorestframework==3.12.4
drf-spectacular==0.18.2
gunicorn==20.1.0
inflection==0.5.1
jsonschema==3.2.0
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
